quicksave

+++ Frituur/Models/Product.cs
@@ -4,9 +4,9 @@
     {
         public int Id { get; set; }
         public string? Name { get; set; }
-        public float? Price { get; set; }
+        public double? Price { get; set; }
         public string? Photo { get; set; }
-        public float? Discount { get; set; }
+        public double? Discount { get; set; }
         public ICollection<Order>? Orders { get; set; }
     }
 }
